requirements.txt
python-telegram-bot==22.0
SQLAlchemy==2.0.36
python-dotenv==1.1.0
matplotlib==3.10.3
PyMySQL==1.1.1
cryptography==44.0.3
Unidecode==1.4.0